requirements.txt
flask==2.3.3
flask-cors==4.0.0
mediapipe==0.10.21
opencv-python==4.9.0.80
numpy==1.26.4
matplotlib==3.8.4
tensorflow-cpu==2.19.0
